bug/fix-search: Changing search type
We were previously doing a term search on a text value.
This has [known limitations](https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-term-query.html]

 This would only work for single word prisons. The text value would be processed during index to split by individual word and lower case whereas the terms query would pass the exact phrase.

So when we searched for 'Cookham wood' it would fail to match the tokens 'cookham' and 'wood'.
Luckily according to the index we could see that a separate  prison_name.keyword was also indexed which indexes the whole thing.

This change moves to using the formatted name and the keyword field as these currently always match.
We should see if drupal would allow us to index establishment id as going forward a keyword index on establishment id would be more robust.